+++ packages/cli/src/commands/config/remove.ts
@@ -25,9 +25,15 @@ export function Remove(_: IContext): CommandModule<{}, Args> {
         demandOption: true,
         choices: fs
           .readdirSync(configsPath)
-          .map((name) =>
-            fs.readdirSync(path.join(configsPath, name)).map((type) => `${name}.${type}`)
-          )
+          .map((name) => {
+            // If no language is detected, default to configs available for typescript
+            const language = Project.detectLanguage() ?? 'typescript';
+            const ttkPath = path.join(configsPath, name);
+
+            return fs.readdirSync(ttkPath)
+              .filter((type) => fs.existsSync(path.join(ttkPath, type, language)))
+              .map((type) => `${name}.${type}`);
+          })
           .flat(),
       });
     },