+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microsoft/teams.cli
 
+## 2.0.0-preview.3
+
+### Patch Changes
+
+- add csharp support to CLI officially, add a2a support via new package, devtools updates
+- Updated dependencies
+  - @microsoft/teams.common@2.0.0-preview.3
+
 ## 2.0.0-preview.2
 
 ### Patch Changes